requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1
django-filter==2.4.0
djangorestframework==3.11.1
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
mixins==0.1.4
packaging==20.8
Pillow==7.1.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-decouple==3.4
pytils==0.3
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
zeep==4.0.0
gunicorn==20.0.4
whitenoise==5.1.0
dj-database-url==0.5.0
djangorestframework-simplejwt
pip install --upgrade stripe

